+++ App.tsx
@@ -1,7 +1,7 @@
 import { StatusBar } from 'expo-status-bar';
 import React from 'react';
 import { StyleSheet, Text, View } from 'react-native';
-import Login from './src/login-page';
+import Login from './src/LoginPage';
 
 export default function App() {
   return (
